moved clear filterts buttons to seperate components
